Tue, Apr 25, 2023, 11:41 AM -04:00

+++ readability/readability.c
@@ -21,10 +21,10 @@ int main(void)
     int sentence_count = count_sentences(text);
     printf("%i sentences\n", sentence_count); // not needed FR
 
-    int L = ((float) letter_count / word_count) * 100;
+    int L = letter_count / word_count * 100;
     printf("L is %i\n", L);
 
-    int S = ((float) sentence_count / word_count) * 100;
+    int S = sentence_count / word_count * 100;
     printf("S is %i\n", S);
 
     int grade_level = 0.0588 * L - 0.296 * S - 15.8;
